# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Tasty Crousty - Smart food ordering and delivery platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ratings:
    post:
      operationId: createRating
      tags: [ratings]
      summary: Submit a rating
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRatingBody"
      responses:
        "201":
          description: Rating created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rating"
  /ratings/restaurant/{restaurantId}:
    get:
      operationId: listRestaurantRatings
      tags: [ratings]
      summary: Get ratings for a restaurant
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ratings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rating"

  # NOTIFICATIONS
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: Get current user notifications
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{notificationId}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /profile:
    get:
      operationId: getProfile
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateProfile
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /addresses:
    get:
      operationId: listAddresses
      tags: [users]
      summary: List user saved addresses
      responses:
        "200":
          description: Addresses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Address"
    post:
      operationId: createAddress
      tags: [users]
      summary: Save a new address
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAddressBody"
      responses:
        "201":
          description: Address created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
  /addresses/{addressId}:
    delete:
      operationId: deleteAddress
      tags: [users]
      summary: Delete a saved address
      parameters:
        - name: addressId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    # AUTH
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
        avatarUrl:
          type: ["string", "null"]
        language:
          type: string
        riskScore:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - language
        - createdAt
    UpdateProfileBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        language:
          type: ["string", "null"]
    Address:
      type: object
      properties:
        id:
          type: integer
        label:
          type: ["string", "null"]
        fullAddress:
          type: string
        landmark:
          type: ["string", "null"]
        floor:
          type: ["string", "null"]
        instructions:
          type: ["string", "null"]
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        isDefault:
          type: boolean
      required:
        - id
        - fullAddress
        - isDefault
    CreateAddressBody:
      type: object
      properties:
        label:
          type: ["string", "null"]
        fullAddress:
          type: string
        landmark:
          type: ["string", "null"]
        floor:
          type: ["string", "null"]
        instructions:
          type: ["string", "null"]
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        isDefault:
          type: boolean
      required:
        - fullAddress

    # CITY / ZONE
...
    Rating:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        restaurantId:
          type: ["integer", "null"]
        driverId:
          type: ["integer", "null"]
        customerId:
          type: integer
        rating:
          type: integer
        comment:
          type: ["string", "null"]
        targetType:
          type: string
          enum: [restaurant, driver]
        createdAt:
          type: string
      required:
        - id
        - orderId
        - customerId
        - rating
        - targetType
        - createdAt
    CreateRatingBody:
      type: object
      properties:
        orderId:
          type: integer
        targetType:
          type: string
          enum: [restaurant, driver]
        targetId:
          type: integer
        rating:
          type: integer
        comment:
          type: ["string", "null"]
      required:
        - orderId
        - targetType
        - targetId
        - rating

    # NOTIFICATIONS
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
        title:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        relatedOrderId:
          type: ["integer", "null"]
        createdAt:
          type: string
      required:
        - id
        - userId
        - type
        - title
        - message
        - isRead
        - createdAt

    # ADMIN
...